Complete rewrite using express and chokidar
Version 0.1.0 was never pushed to GitHub, but was published on
npm. This is a complete rewrite which uses third party packages
to reduce the amount of first party code and subsequent bugs.

Version 0.2.0 isn’t feature complete but is stable enough for use
with my personal blog via a Nunjucks plugin. This will be
published as a separate package soon.